Backend/requirements.txt
openai
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
langchain-community
pypdf